PresentationLinksCard component
