+++ src/data/mock/systemIntake.ts
@@ -161,7 +161,7 @@ export const documents: SystemIntakeDocument[] = [
   }
 ];
 
-const grbPresentationLinks: SystemIntake['grbPresentationLinks'] = {
+export const grbPresentationLinks: SystemIntake['grbPresentationLinks'] = {
   __typename: 'SystemIntakeGRBPresentationLinks',
   systemIntakeID: '5af245bc-fc54-4677-bab1-1b3e798bb43c',
   createdAt: '2025-01-17T04:20:32.122480967Z',
